# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: WTS – 모의투자 Web Trading System API
  version: 0.0.1-SNAPSHOT
  description: '모의투자 Web Trading System API.


    클라이언트는 각 서비스를 직접 호출하지 않고 Gateway(http://localhost:8080)로만 접근한다.

    토큰은 `POST /api/users/mock-login` 으로 발급받는다.


    주문 API는 `Idempotency-Key` 헤더가 필수다. 같은 키로 같은 내용을

    재요청하면 주문이 새로 생기지 않고 기존 결과가 돌아온다.


    모든 응답에 `X-Trace-Id` 헤더가 붙는다. 에러 본문의 `traceId`와 같은 값이다.

    '
servers:
- url: http://localhost:8080
  description: Gateway
security:
- bearerAuth: []
...
paths:
  /api/market/stocks:
    get:
      tags:
      - 종목
      summary: 종목 목록·검색
      description: keyword를 주면 종목코드 또는 종목명 부분일치로 검색한다. 생략하면 전체 목록이다.
      operationId: list
      parameters:
      - name: keyword
        in: query
        description: 종목코드 또는 종목명 일부. 생략하면 전체
        required: false
        schema:
          type: string
        example: 삼성
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockResponse'
      security: []
      x-error-codes: []
  /api/market/stocks/{symbol}:
    get:
      tags:
      - 종목
      summary: 종목 상세
      description: 종목코드로 종목 하나를 조회한다.
      operationId: detail
      parameters:
      - name: symbol
        in: path
        description: 종목코드 (6자리)
        required: true
        schema:
          type: string
        example: 005930
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockResponse'
        '404':
          description: 'Not Found


            - `SYMBOL_NOT_FOUND` — 종목을 찾을 수 없습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
      x-error-codes:
      - SYMBOL_NOT_FOUND
  /api/market/stocks/{symbol}/candles:
    get:
      tags:
      - 차트
      summary: 분봉 조회
      description: 최근 분봉을 openTime 오름차순으로 돌려준다. 마지막 원소는 아직 확정되지 않은 현재 봉일 수 있다. 현재 1분봉만 지원한다.
      operationId: candles
      parameters:
      - name: symbol
        in: path
        description: 종목코드 (6자리)
        required: true
        schema:
          type: string
        example: 005930
      - name: interval
        in: query
        description: 봉 주기. 현재 1m만 지원
        required: false
        schema:
          type: string
          default: 1m
        example: 1m
      - name: limit
        in: query
        description: 최근 몇 개를 받을지. 1 ~ 1000
        required: false
        schema:
          type: integer
          format: int32
          default: 120
        example: 120
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CandleResponse'
        '400':
          description: 'Bad Request


            - `VALIDATION_FAILED` — 요청 값이 올바르지 않습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 'Not Found


            - `SYMBOL_NOT_FOUND` — 종목을 찾을 수 없습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
      x-error-codes:
      - VALIDATION_FAILED
      - SYMBOL_NOT_FOUND
  /api/market/stocks/{symbol}/orderbook:
    get:
      tags:
      - 시세
      summary: 호가 조회
      description: 최신 매도·매수 호가를 돌려준다. 매도호가는 낮은 가격부터, 매수호가는 높은 가격부터 정렬한다.
      operationId: orderBook
      parameters:
      - name: symbol
        in: path
        description: 종목코드 (6자리)
        required: true
        schema:
          type: string
        example: 005930
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderBookResponse'
        '404':
          description: 'Not Found


            - `SYMBOL_NOT_FOUND` — 종목을 찾을 수 없습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 'Service Unavailable


            - `MARKET_PRICE_UNAVAILABLE` — 현재가를 조회할 수 없습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
      x-error-codes:
      - SYMBOL_NOT_FOUND
      - MARKET_PRICE_UNAVAILABLE
  /api/market/stocks/{symbol}/price:
    get:
      tags:
      - 시세
      summary: 현재가 조회
      description: 최신 현재가를 돌려준다. 필드 구성은 WebSocket PRICE 메시지와 같다. 실시간 갱신이 필요하면 WebSocket(/ws/market)을 구독한다.
        아직 시세가 한 번도 들어오지 않았으면 503이다.
      operationId: price
      parameters:
      - name: symbol
        in: path
        description: 종목코드 (6자리)
        required: true
        schema:
          type: string
        example: 005930
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceResponse'
        '404':
          description: 'Not Found


            - `SYMBOL_NOT_FOUND` — 종목을 찾을 수 없습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 'Service Unavailable


            - `MARKET_PRICE_UNAVAILABLE` — 현재가를 조회할 수 없습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
      x-error-codes:
      - SYMBOL_NOT_FOUND
      - MARKET_PRICE_UNAVAILABLE
...
components:
  schemas:
...
    CandleResponse:
      type: object
      properties:
        openTime:
          type: string
          format: date-time
          description: 봉 시작 시각 (UTC). 해당 분의 0초
          example: '2026-09-22T00:01:00Z'
        open:
          type: number
          description: 시가 (원)
          example: 80300
        high:
          type: number
          description: 고가 (원)
          example: 80500
        low:
          type: number
          description: 저가 (원)
          example: 79000
        close:
          type: number
          description: 종가 (원)
          example: 79800
        volume:
          type: integer
          format: int64
          description: 해당 1분 동안의 거래량 (주)
          example: 158222
    ErrorResponse:
      properties:
        code:
          type: string
          description: 에러 코드
          example: INSUFFICIENT_BALANCE
        message:
          type: string
          description: 사람이 읽을 수 있는 메시지
          example: 주문 가능한 예수금이 부족합니다.
        traceId:
          type: string
          description: 추적 ID. 응답 헤더 X-Trace-Id와 같다
          example: effc7d36-ebd7-46ae-9ad8-9ea215d500b5
        timestamp:
          type: string
          format: date-time
          description: 발생 시각 (UTC)
          example: '2026-09-22T08:12:44.101Z'
...
    OrderBookLevel:
      type: object
      properties:
        price:
          type: number
          description: 호가 (원)
          example: 79900
        quantity:
          type: integer
          format: int64
          description: 잔량 (주)
          example: 835
    OrderBookResponse:
      type: object
      properties:
        symbol:
          type: string
          description: 종목코드
          example: 005930
        asks:
          type: array
          description: 매도호가. 최우선(가장 낮은 가격)부터
          items:
            $ref: '#/components/schemas/OrderBookLevel'
        bids:
          type: array
          description: 매수호가. 최우선(가장 높은 가격)부터
          items:
            $ref: '#/components/schemas/OrderBookLevel'
        timestamp:
          type: string
          format: date-time
          description: 호가 시각 (UTC, ISO-8601)
          example: '2026-09-22T08:12:44.101Z'
...
    PriceResponse:
      type: object
      properties:
        symbol:
          type: string
          description: 종목코드
          example: 005930
        price:
          type: number
          description: 현재가 (원)
          example: 79800
        previousClose:
          type: number
          description: 전일 종가 (원)
          example: 80000
        change:
          type: number
          description: 전일 대비 (원) = price - previousClose
          example: -200
        changeRate:
          type: number
          description: 전일 대비 등락률 (%), 소수 둘째 자리
          example: -0.25
        volume:
          type: integer
          format: int64
          description: 당일 누적 거래량 (주). 틱 증분이 아니다
          example: 159379
        timestamp:
          type: string
          format: date-time
          description: 시세 시각 (UTC, ISO-8601)
          example: '2026-09-22T08:12:44.101Z'
    StockResponse:
      type: object
      properties:
        symbol:
          type: string
          description: 종목코드 (6자리)
          example: 005930
        name:
          type: string
          description: 종목명
          example: 삼성전자
        market:
          type: string
          description: 시장 구분
          example: KOSPI
...
  securitySchemes:
    bearerAuth:
      type: http
      description: '`POST /api/users/mock-login` 이 발급한 accessToken'
      scheme: bearer
...